+++ internal/serve/serve.go
@@ -84,7 +84,7 @@ type ServeOptions struct {
 	DisableMFA                      bool
 	DisableReCAPTCHA                bool
 	PasswordValidator               *authUtils.PasswordValidator
-	EnableScheduler                 bool
+	EnableScheduler                 bool // Deprecated: Use EventBrokerType=SCHEDULER instead.
 	tenantManager                   tenant.ManagerInterface
 	DistributionAccountService      services.DistributionAccountServiceInterface
 	DistAccEncryptionPassphrase     string
@@ -106,7 +106,10 @@ func (opts *ServeOptions) SetupDependencies() error {
 	httperror.SetDefaultReportErrorFunc(opts.CrashTrackerClient.LogAndReportErrors)
 
 	// Setup Multi-Tenant Database when enabled
-	opts.tenantManager = tenant.NewManager(tenant.WithDatabase(opts.AdminDBConnectionPool))
+	opts.tenantManager = tenant.NewManager(
+		tenant.WithDatabase(opts.AdminDBConnectionPool),
+		tenant.WithSingleTenantMode(opts.SingleTenantMode),
+	)
 
 	var err error
 	opts.Models, err = data.NewModels(opts.MtnDBConnectionPool)
@@ -143,7 +146,7 @@ func (opts *ServeOptions) ValidateSecurity() error {
 		if opts.DisableMFA {
 			return fmt.Errorf("MFA cannot be disabled in pubnet")
 		} else if opts.DisableReCAPTCHA {
-			return fmt.Errorf("reCAPTCHA cannot be disabled in pubnet")
+			log.Warnf("reCAPTCHA is disabled in pubnet. This might reduce security!")
 		}
 	}